requirements.txt
bidict==0.23.1
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
idna==3.8
inquirerpy==0.3.4
packaging==24.1
pfzy==0.3.4
prompt_toolkit==3.0.47
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
simple-websocket==1.0.0
uptime_kuma_api==1.2.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
